% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_breaks_class.R
\name{ltm_add_runs}
\alias{ltm_add_runs}
\title{Add break-detection runs to a \code{ts_breaks} container}
\usage{
ltm_add_runs(ts_breaks_obj, ...)
}
\arguments{
\item{ts_breaks_obj}{An object of class \code{ts_breaks}.}

\item{...}{One or more objects of class \code{ts_breaks_run}, typically returned
by the break-detection wrapper functions.}
}
\value{
The updated \code{ts_breaks} object.
}
\description{
Adds one or more \code{ts_breaks_run} objects to a \code{ts_breaks} container. Runs are
grouped by their \code{method} field and assigned sequential run identifiers such
as \code{"run-01"}.
}
\seealso{
Other break result helpers: 
\code{\link{as.data.frame.ts_breaks}()},
\code{\link{ltm_get_algorithms}()},
\code{\link{ltm_get_run_details}()},
\code{\link{ltm_get_runs}()},
\code{\link{ltm_summarize_break_df}()},
\code{\link{ltm_ts_breaks}()}
}
\concept{break result helpers}
