% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_gee_data.R
\name{ltm_ee_feature_collection_to_df}
\alias{ltm_ee_feature_collection_to_df}
\title{Convert an Earth Engine FeatureCollection to a data frame}
\usage{
ltm_ee_feature_collection_to_df(x_fc, max_features = 10000L)
}
\arguments{
\item{x_fc}{An `ee$FeatureCollection` object.}

\item{max_features}{Integer. Maximum number of features allowed to be
retrieved. Defaults to 10000.}
}
\value{
A data frame where each row corresponds to a feature and columns correspond
  to feature properties.
}
\description{
Retrieves a FeatureCollection from Google Earth Engine, checks its size
against a maximum threshold, and converts its features into a data frame.
}
\keyword{internal}
