% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_gee_data.R
\name{ltm_ee_features_to_df}
\alias{ltm_ee_features_to_df}
\title{Convert a list of Earth Engine features to a data frame}
\usage{
ltm_ee_features_to_df(features)
}
\arguments{
\item{features}{A list of Earth Engine feature objects.}
}
\value{
A data frame where each row corresponds to a feature and columns correspond
to feature properties. Returns an empty data frame if input is \code{NULL} or empty.
}
\description{
Converts a list of Earth Engine feature objects into a data frame by
extracting their properties. Ensures that a \code{system.id} field is present
when available and standardizes column names.
}
