% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_breaks_class.R
\name{ltm_get_algorithms}
\alias{ltm_get_algorithms}
\title{List algorithms stored in a \code{ts_breaks} container}
\usage{
ltm_get_algorithms(ts_breaks_obj)
}
\arguments{
\item{ts_breaks_obj}{An object of class \code{ts_breaks}.}
}
\value{
Character vector of algorithm names.
}
\description{
List algorithms stored in a \code{ts_breaks} container
}
\seealso{
Other break result helpers: 
\code{\link{as.data.frame.ts_breaks}()},
\code{\link{ltm_add_runs}()},
\code{\link{ltm_get_run_details}()},
\code{\link{ltm_get_runs}()},
\code{\link{ltm_summarize_break_df}()},
\code{\link{ltm_ts_breaks}()}
}
\concept{break result helpers}
