% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_breaks_class.R
\name{ltm_get_run_details}
\alias{ltm_get_run_details}
\title{Retrieve one break-detection run}
\usage{
ltm_get_run_details(ts_breaks_obj, algorithm_name, run_id)
}
\arguments{
\item{ts_breaks_obj}{An object of class \code{ts_breaks}.}

\item{algorithm_name}{Character scalar algorithm name present in
\code{ts_breaks_obj}.}

\item{run_id}{Character scalar run identifier for \code{algorithm_name}.}
}
\value{
A list representing one stored \code{ts_breaks_run}.
}
\description{
Retrieve one break-detection run
}
\seealso{
Other break result helpers: 
\code{\link{as.data.frame.ts_breaks}()},
\code{\link{ltm_add_runs}()},
\code{\link{ltm_get_algorithms}()},
\code{\link{ltm_get_runs}()},
\code{\link{ltm_summarize_break_df}()},
\code{\link{ltm_ts_breaks}()}
}
\concept{break result helpers}
