% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_breaks_class.R
\name{ltm_get_runs}
\alias{ltm_get_runs}
\title{List run identifiers for an algorithm}
\usage{
ltm_get_runs(ts_breaks_obj, algorithm_name)
}
\arguments{
\item{ts_breaks_obj}{An object of class \code{ts_breaks}.}

\item{algorithm_name}{Character scalar algorithm name present in
\code{ts_breaks_obj}.}
}
\value{
Character vector of run identifiers for \code{algorithm_name}.
}
\description{
List run identifiers for an algorithm
}
\seealso{
Other break result helpers: 
\code{\link{as.data.frame.ts_breaks}()},
\code{\link{ltm_add_runs}()},
\code{\link{ltm_get_algorithms}()},
\code{\link{ltm_get_run_details}()},
\code{\link{ltm_summarize_break_df}()},
\code{\link{ltm_ts_breaks}()}
}
\concept{break result helpers}
