% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_aux_funs.R
\name{ltm_spidf_to_zoo}
\alias{ltm_spidf_to_zoo}
\title{Convert an \code{spidf} column to a \code{zoo} time series}
\usage{
ltm_spidf_to_zoo(spidf_obj, yname)
}
\arguments{
\item{spidf_obj}{An object of class \code{spidf}.}

\item{yname}{Character scalar naming the column to use as the time-series
values.}
}
\value{
A \code{zoo} object indexed by \code{spidf_obj$ti}.
}
\description{
Creates a \code{\link[zoo:zoo]{zoo::zoo()}} object using the selected value column and the \code{ti}
dates stored in an \code{spidf} object.
}
\seealso{
Other time-series conversion helpers: 
\code{\link{ltm_copy_ts}()},
\code{\link{ltm_days_between}()},
\code{\link{ltm_spidf_to_ts}()}
}
\concept{time-series conversion helpers}
