% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_breaks_class.R
\name{ltm_ts_breaks}
\alias{ltm_ts_breaks}
\title{Create a break-detection result container}
\usage{
ltm_ts_breaks(spidf_ts)
}
\arguments{
\item{spidf_ts}{An object of class \code{spidf}.}
}
\value{
An object of class \code{ts_breaks} with elements \code{spidf_ts} and
\code{algorithms}.
}
\description{
Initializes a \code{ts_breaks} object for an \code{spidf} time series. The container
stores the original time series and a nested list of break-detection runs
grouped by algorithm name.
}
\seealso{
Other break result helpers: 
\code{\link{as.data.frame.ts_breaks}()},
\code{\link{ltm_add_runs}()},
\code{\link{ltm_get_algorithms}()},
\code{\link{ltm_get_run_details}()},
\code{\link{ltm_get_runs}()},
\code{\link{ltm_summarize_break_df}()}
}
\concept{break result helpers}
