% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTM_shiny_app.R
\name{run_ltm_app}
\alias{run_ltm_app}
\title{Run the LargeTreeMonitoring Shiny app}
\usage{
run_ltm_app(...)
}
\arguments{
\item{...}{Arguments passed to shiny::runApp()}
}
\description{
Runs the shiny app GUI to get data for a specific point. This point can be
provided by a dropdown list from a CSV file set by the JSON parameters file.
}
